refactor(stelae): move the staging arithmetic behind a state-kind hook
The 2026-09-01 ruling on this plan's escalation, in three parts.

`DriverProfile` grows `is_state_kind` — pure kind classification, the same
nature as `epoch_kinds` and `is_inheritable`, and the one question the
staging arithmetic asks that the other four cannot answer: `is_inheritable`
is true for `blocks` and `indexes` and false for both a tip shard and
`digests`, so no combination of it separates the layers that must be summed
from the one that must not.

With it, `StagingPeak`, `staging_peak`, `staging_need` and the publish
`preflight` move to `stelae-driver`, generic over the trait. They are
re-exported at their old paths in `registry`; the profile the driver needs
is now the caller's to supply.

`Origin`'s `network_magic` becomes `dataset_id`, field and wire. A
dataset-shaped name has no place in a profile-generic crate's record schema,
and this is the cheap moment: the record is host-local resumption state, not
published protocol. A record written by an older binary no longer matches, so
a publish in flight across the upgrade restarts rather than resumes.
Inscription digests are untouched.
